chore: move httpInterceptor from CMF
and start refactoring the http client
